services/correction-service/requirements.txt
fastapi
uvicorn[standard]
spacy
google-generativeai